Ctypes based Windows implementation of pyopen8055.py.
At this point a Windows PC cannot switch to the second
configuration because Microsoft apparently doesn't support
multiple configurations for generic HID devices.

For now pyopen8055 can work with real K8055 and K8055N
cards in their native protocol and use an Open8055 like
a K8055. Not good, but I will find a solution for that
later.